cleanup effects and its files
