+++ Src/Projects/manager_PostProcessing/posteffectbase.cxx
@@ -10,6 +10,7 @@ Licensed under The "New" BSD License - https://github.com/Neill3d/OpenMoBu/blob/
 
 //--- Class declaration
 #include "posteffectbase.h"
+#include "postpersistentdata.h"
 
 //
 extern void LOGE(const char* pFormatString, ...);
