+++ Src/Projects/manager_PostProcessing/posteffectchain.h
@@ -1,0 +1,107 @@
+#pragma once
+
+// postprocessing_effectChain
+/*
+Sergei <Neill3d> Solokhin 2018-2024
+
+GitHub page - https://github.com/Neill3d/OpenMoBu
+Licensed under The "New" BSD License - https://github.com/Neill3d/OpenMoBu/blob/master/LICENSE
+*/
+
+//--- SDK include
+#include <fbsdk/fbsdk.h>
+
+#include "GL/glew.h"
+
+#include "graphics_framebuffer.h"
+#include "postpersistentdata.h"
+#include "posteffectbase.h"
+
+#include "glslShader.h"
+#include "Framebuffer.h"
+
+#include <memory>
+#include <bitset>
+
+// forward
+class PostEffectBuffers;
+
+/// <summary>
+/// chain of post processing effects, apply effects in an order
+/// </summary>
+class PostEffectChain
+{
+
+public:
+	//! a constructor
+	PostEffectChain();
+	//! a destructor
+	~PostEffectChain();
+
+	void ChangeContext();
+	/// w,h - local buffer size for processing, pCamera - current pane camera for processing
+	bool Prep(PostPersistentData *pData, int w, int h, FBCamera *pCamera);
+
+	bool BeginFrame(PostEffectBuffers* buffers);
+	bool Process(PostEffectBuffers* buffers, double time);
+
+	bool IsCompressedDataReady()
+	{
+		return mIsCompressedDataReady;
+	}
+
+protected:
+
+	FBSystem								mSystem;
+	HdlFBPlugTemplate<PostPersistentData>	mSettings;
+	FBCamera								*mLastCamera;
+
+	// instances of each effect
+	std::unique_ptr<PostEffectBase>		mFishEye;
+	std::unique_ptr<PostEffectBase>		mColor;
+	std::unique_ptr<PostEffectBase>		mVignetting;
+	std::unique_ptr<PostEffectBase>		mFilmGrain;
+	std::unique_ptr<PostEffectBase>		mLensFlare;
+	std::unique_ptr<PostEffectBase>		mSSAO;
+	std::unique_ptr<PostEffectBase>		mDOF;
+	std::unique_ptr<PostEffectBase>		mDisplacement;
+	std::unique_ptr<PostEffectBase>		mMotionBlur;
+
+	// shared shaders
+	
+	std::unique_ptr<GLSLShader>			mShaderDepthLinearize;	//!< linearize depth for other filters (DOF, SSAO, Bilateral Blur, etc.)
+	std::unique_ptr<GLSLShader>			mShaderBlur;	//!< needed for SSAO
+	std::unique_ptr<GLSLShader>			mShaderImageBlur; //!< for masking
+	std::unique_ptr<GLSLShader>			mShaderMix;		//!< needed for SSAO
+	std::unique_ptr<GLSLShader>			mShaderDownscale;
+
+	// order execution chain
+	std::vector<PostEffectBase*>		mChain;
+
+	GLint							mLocDepthLinearizeClipInfo;
+	GLint							mLocBlurSharpness;
+	GLint							mLocBlurRes;
+	GLint							mLocImageBlurScale;
+
+	bool							mNeedReloadShaders;
+	bool							mIsCompressedDataReady;
+	double							mLastCompressTime;
+
+	PostEffectBase *ShaderFactory(const int type, const char *shadersLocation);
+
+	bool LoadShaders();
+	void FreeShaders();
+
+private:
+	static bool CheckShadersPath(const char* path);
+
+	/// <summary>
+	/// return true if a global masking active and any of effects uses a masking
+	/// </summary>
+	bool HasMaskUsedByEffect() const;
+
+	/// <summary>
+	/// return true if there is any model with FXMaskingShader, so that we could render it into a masking texture
+	/// </summary>
+	bool HasAnyMaskedObject() const;
+};
