+++ Src/Projects/manager_PostProcessing/posteffectlensflare.cpp
@@ -1,13 +1,14 @@
 
-// postprocessing_effectSSAO.cpp
+// posteffectlensflare.cpp
 /*
-Sergei <Neill3d> Solokhin 2018
+Sergei <Neill3d> Solokhin 2018-2024
 
 GitHub page - https://github.com/Neill3d/OpenMoBu
 Licensed under The "New" BSD License - https://github.com/Neill3d/OpenMoBu/blob/master/LICENSE
 */
 
-#include "postprocessing_effectLensFlare.h"
+#include "posteffectlensflare.h"
+#include "postpersistentdata.h"
 
 #define _USE_MATH_DEFINES
 #include <math.h>
