+++ Src/Projects/manager_PostProcessing/posteffectmotionblur.h
@@ -1,20 +1,20 @@
 
 #pragma once
 
-// postprocessing_effectMotionBlur
+// posteffectmotionblur
 /*
-Sergei <Neill3d> Solokhin 2018
+Sergei <Neill3d> Solokhin 2018-2024
 
 GitHub page - https://github.com/Neill3d/OpenMoBu
 Licensed under The "New" BSD License - https://github.com/Neill3d/OpenMoBu/blob/master/LICENSE
 */
 
-#include "postprocessing_effectChain.h"
+#include "posteffectbase.h"
 #include <random>
 
-///////////////////////////////////////////////////////////////////////////////////////
-// PostEffectMotionBlur
-
+/// <summary>
+/// camera motion blur post processing effect
+/// </summary>
 struct PostEffectMotionBlur : public PostEffectBase, public CommonEffectUniforms
 {
 public:
@@ -34,7 +34,7 @@ public:
 	const char* GetEnableMaskPropertyName() const override { return "Motion Blur Use Masking"; }
 
 	virtual bool PrepUniforms(const int shaderIndex) override;
-	virtual bool CollectUIValues(PostPersistentData *pData, int w, int h, FBCamera *pCamera) override;
+	virtual bool CollectUIValues(PostPersistentData* pData, int w, int h, FBCamera* pCamera) override;
 
 	virtual void Bind() override;
 	virtual void UnBind() override;
